refactor:  change grading-job-actions to grading-job-options
